Refactor the boxes reset, resync, transfer, reboot methods [skip ci]
This commit changes the way the above functions work.
We have recently changed the backend API to be more restful.

Instead of calling a POST against the reboot / reset / resync
endpoints, you now send in an action param. For instance,

PATCH update action = resync would resync the device.

To transfer a device, you now update the device with the
transfer_to param. For example:

PATCH update transfer_to = 123

Boxes can now only be transferred WITHIN an account, not
to another user (for billing reasons).

Finally, the location transfer methods have been updated
as per the box updates.

Finally part II. To archive a location, you now update the
state of the location and don't need to send a dedicated flag